refactor: allow builders to have custom defaults
